Add Code, apply gRPC service changes to drop tonic, drop CodeExt
