+++ mocktail/Cargo.toml
@@ -31,7 +31,6 @@ serde_json = "1"
 thiserror = "2"
 tokio = "1"
 tokio-stream = "0"
-tonic = "0.12"
 tracing = "0"
 url = "2"
 uuid = { version = "1.16.0", features = ["fast-rng", "v7"] }
