+++ mocktail/src/service/grpc.rs
@@ -9,10 +9,11 @@ use http_body_util::{BodyExt, StreamBody};
 use hyper::{body::Incoming, service::Service};
 use tokio::sync::mpsc;
 use tokio_stream::wrappers::ReceiverStream;
-use tonic::body::BoxBody;
 use tracing::debug;
 
-use crate::{request::Request, server::MockServerState};
+use crate::{request::Request, server::MockServerState, service::http::empty, Code};
+
+use super::http::BoxBody;
 
 /// Mock gRPC service.
 #[derive(Debug, Clone)]
@@ -40,7 +41,7 @@ impl Service<http::Request<Incoming>> for GrpcMockService {
                 return Ok(http::Response::builder()
                     .status(http::StatusCode::METHOD_NOT_ALLOWED)
                     .header("Allow", "POST")
-                    .body(tonic::body::empty_body())
+                    .body(empty())
                     .unwrap());
             }
             let content_type = req.headers().get("content-type");
@@ -52,7 +53,7 @@ impl Service<http::Request<Incoming>> for GrpcMockService {
                 return Ok(http::Response::builder()
                     .status(http::StatusCode::UNSUPPORTED_MEDIA_TYPE)
                     .header("Accept-Post", "application/grpc")
-                    .body(tonic::body::empty_body())
+                    .body(empty())
                     .unwrap());
             }
 
@@ -61,7 +62,7 @@ impl Service<http::Request<Incoming>> for GrpcMockService {
 
             // Create response stream
             let (response_tx, response_rx) =
-                mpsc::channel::<Result<Frame<Bytes>, tonic::Status>>(32);
+                mpsc::channel::<Result<Frame<Bytes>, hyper::Error>>(32);
             let response_stream = ReceiverStream::new(response_rx);
             let response_body = BoxBody::new(StreamBody::new(response_stream));
             let response = http::Response::builder()
@@ -122,7 +123,7 @@ impl Service<http::Request<Incoming>> for GrpcMockService {
 
 fn mock_not_found_trailer() -> HeaderMap {
     let mut headers = HeaderMap::new();
-    headers.insert("grpc-status", (tonic::Code::NotFound as i32).into());
+    headers.insert("grpc-status", (Code::NotFound as i32).into());
     headers.insert("grpc-message", HeaderValue::from_static("mock not found"));
     headers
 }
